Testing: Fix wrong string compare in testing framework. Since the line content is stored _without_ a trailing null terminator, memcmp must be used instead of strcmp.
Testing: Fix buffer tests. Since free doesn't change the pointer itself,
it can't be checked if a free was successful.

Testing: Fix memory leaks by using a goto-cleanup construct. It is on
the one hand headache-making to use goto but on the other side is it
only meant for clean up and to avoid making the tests WET (Opposite of
Dont-Repeat-Yourself) where each test-failed-branch has to do the same.
Readability improvement, i swear (:fingers-crossed:).